feat(security): Authorization

+++ concerts/views.py
@@ -1,9 +1,10 @@
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
+from django.contrib.auth.mixins import LoginRequiredMixin
 from django.core.exceptions import ValidationError
 from django.db.models import Count, Q
 from django.db.models.functions import Lower
-from django.http import HttpResponseRedirect, JsonResponse
+from django.http import HttpResponseForbidden, HttpResponseRedirect, JsonResponse
 from django.shortcuts import get_object_or_404, redirect, render, reverse
 from django.urls import reverse_lazy
 from django.views import View
@@ -93,9 +94,7 @@ def unattend_concert(request, pk, next=None):
     concert = get_object_or_404(Concert, pk=pk)
     concert.attendees.remove(request.user)
 
-    if next == "concert-list":
-        return redirect("concerts:concert-list")
-    elif next == "user-detail":
+    if next == "user-detail":
         return redirect("users:detail", request.user.username)
     else:
         return redirect("concerts:concert-list")
@@ -130,7 +129,7 @@ class ArtistDetailView(DetailView):
             return render(self.request, self.template_name, context)
 
 
-class ArtistCreateView(CreateView):
+class ArtistCreateView(LoginRequiredMixin, CreateView):
     model = Artist
     template_name = "concerts/artist_create.html"
     form_class = ArtistForm
@@ -142,7 +141,7 @@ class ArtistCreateView(CreateView):
         artist_instance = form.save()
 
         # Check if the request is AJAX
-        if self.request.headers.get("x-requested-with") == "XMLHttpRequest":
+        if self.request.headers.get("X-Requested-With") == "XMLHttpRequest":
             data = {
                 "status": "success",
                 "artist": artist_instance.name,
@@ -157,7 +156,7 @@ class ArtistCreateView(CreateView):
 
     def form_invalid(self, form):
         # Check if the request is AJAX
-        if self.request.headers.get("x-requested-with") == "XMLHttpRequest":
+        if self.request.headers.get("X-Requested-With") == "XMLHttpRequest":
             errors = {field: error[0] for field, error in form.errors.items()}
             return JsonResponse({"status": "error", "errors": errors}, status=400)
 
@@ -184,7 +183,7 @@ class VenueDetailView(DetailView):
         return context
 
 
-class VenueCreateView(CreateView):
+class VenueCreateView(LoginRequiredMixin, CreateView):
     model = Venue
     template_name = "concerts/venue_create.html"
     form_class = VenueForm
@@ -196,7 +195,7 @@ class VenueCreateView(CreateView):
         venue = form.save()
 
         # Check if the request is AJAX
-        if self.request.headers.get("x-requested-with") == "XMLHttpRequest":
+        if self.request.headers.get("X-Requested-With") == "XMLHttpRequest":
             return JsonResponse(
                 {
                     "status": "success",
@@ -213,7 +212,7 @@ class VenueCreateView(CreateView):
         errors = {field: error[0] for field, error in form.errors.items()}
 
         # Check if the request is AJAX
-        if self.request.headers.get("x-requested-with") == "XMLHttpRequest":
+        if self.request.headers.get("X-Requested-With") == "XMLHttpRequest":
             return JsonResponse({"status": "error", "errors": errors}, status=400)
 
         # For non-AJAX requests:
@@ -280,7 +279,11 @@ class ConcertDetailView(DetailView):
 
     def get_context_data(self, *args, **kwargs):
         context = super().get_context_data(**kwargs)
-        context["user_concerts"] = Concert.objects.filter(attendees=self.request.user).values_list("id", flat=True)
+        if self.request.user.is_authenticated:
+            context["user_concerts"] = Concert.objects.filter(attendees=self.request.user).values_list("id", flat=True)
+        else:
+            context["user_concerts"] = []
+
         return context
 
     def render_to_response(self, context, **response_kwargs):
@@ -298,7 +301,7 @@ class ConcertDetailView(DetailView):
         return super().render_to_response(context, **response_kwargs)
 
 
-class ConcertCreateView(CreateView):
+class ConcertCreateView(LoginRequiredMixin, CreateView):
     model = Concert
     template_name = "concerts/concert_create.html"
     form_class = ConcertForm
@@ -356,7 +359,7 @@ class ConcertCreateView(CreateView):
         return super().form_invalid(form)
 
 
-class ConcertReviewCreateView(CreateView):
+class ConcertReviewCreateView(LoginRequiredMixin, CreateView):
     model = ConcertReview
     form_class = ConcertReviewForm
     template_name = "users/user_detail.html"
@@ -390,7 +393,7 @@ class ConcertReviewCreateView(CreateView):
         errors = {field: error[0] for field, error in form.errors.items()}
 
         # Check if the request is AJAX
-        if self.request.headers.get("x-requested-with") == "XMLHttpRequest":
+        if self.request.headers.get("X-Requested-With") == "XMLHttpRequest":
             return JsonResponse({"status": "error", "errors": errors}, status=400)
 
         # For non-AJAX requests:
@@ -398,7 +401,7 @@ class ConcertReviewCreateView(CreateView):
         return self.render_to_response(self.get_context_data(form=form))
 
 
-class ConcertReviewDetailView(View):
+class ConcertReviewDetailView(LoginRequiredMixin, View):
     def get(self, request, *args, **kwargs):
         review_id = kwargs.get("review_id")
         try:
@@ -413,32 +416,39 @@ class ConcertReviewDetailView(View):
             return JsonResponse({"error": "Review not found"}, status=404)
 
 
-class ConcertReviewUpdateView(View):
+class ConcertReviewUpdateView(LoginRequiredMixin, View):
     def post(self, request, *args, **kwargs):
         review_id = self.kwargs.get("review_id")
         review = get_object_or_404(ConcertReview, id=review_id)
+
+        # Check if the current user is the author of the review
+        if review.user != self.request.user:
+            return HttpResponseForbidden("You don't have permission to edit this review.")
 
         note = request.POST.get("note")
         rating = request.POST.get("rating")
 
         review.note = note
         review.rating = rating
-        review.save()
 
         try:
-            # Attempt to save. This may fail due to model field validations.
             review.full_clean()  # This will run all model field validations
             review.save()
         except ValidationError as e:
-            # Catch validation errors and return them
             return JsonResponse({"success": False, "errors": e.message_dict}, status=400)
 
         return JsonResponse({"success": True, "note": review.note, "rating": review.rating, "reviewId": review_id})
 
 
-class ConcertReviewDeleteView(View):
+class ConcertReviewDeleteView(LoginRequiredMixin, View):
     def delete(self, request, *args, **kwargs):
         review_id = kwargs.get("review_id")
+        review = get_object_or_404(ConcertReview, id=review_id)
+
+        # Check if the current user is the author of the review
+        if review.user != self.request.user:
+            return HttpResponseForbidden("You don't have permission to edit this review.")
+
         try:
             review = ConcertReview.objects.get(pk=review_id)
             review.delete()
